pyproject.toml
[project]
name = "guideforcoach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "e>=1.4.5",
    "ipykernel>=7.2.0",
    "joblib>=1.5.3",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.9",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "tavily-python>=0.7.21",
]

#[tool.setuptools]
#packages = ["src"]

#[tool.setuptools.package-dir]
#packages.find = { where = ["src"] }

#this last line maps the import name to the folder name. basically it helps in executing src folder in jupyter notebook
#else we want to manually do via os